Configure a WebSecurityConfigurerAdapter to use the new stuff
It sets the AuthenticationManager to use the UserDetailsService
added before.
It applies the JwtRequestFilter before regular user+pass
authentication.
It is configured to require authentication on every API endpoint,
save for the '/auth' route.
Since we are using JWT, the session looks differently in the
server's memory (SessionCreationPolicy is set to STATELESS)