Update User to reflect linking with Events
Users now have a Set<Event> named 'eventsJoined'.
This is the counterpart in the many-to-many relationship.
(We decided that the Events own Users, as far as annotations go)
The .joinEvent() and .leaveEvent() make a double link between
the User and Event instances. Double linking also applies in
setter...
(As far as linking in Java is concerned, Users are responsible
for events)